+++ docs/api/changelog.rst
@@ -3,6 +3,13 @@
 Changelog
 ---------
 
+.. _release-0.5.4:
+
+0.5.4 - TBD
+    * Will now check return types for methods and functions more thorouhgly
+    * Will throw errors if a type guard is used with a concrete annotation that uses
+      a type var (mypy plugin system is limited in a way that makes this impossible to implement)
+
 .. _release-0.5.3:
 
 0.5.3 - 25 May 2024